# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Storage Application API",
        "description": "Storage API",
        "contact": {
            "name": "Min",
            "url": "https://www.facebook.com/Min9802/",
            "email": "hotro.minservice@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/storage/client/file/list": {
            "get": {
                "tags": [
                    "Storage Api file"
                ],
                "summary": "get all file",
                "description": "use token",
                "operationId": "get all file",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/FileSystem"
                                        }
                                    ]
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "jx3m8NPBgy.png",
                                        "type": "image",
                                        "path": "public/image/jx3m8NPBgy.png",
                                        "status": 0,
                                        "created_at": "2022-09-29T17:03:15.000000Z",
                                        "updated_at": "2022-09-29T17:13:16.000000Z",
                                        "deleted_at": null,
                                        "url": "http://domain.com/storage/image/jx3m8NPBgy.png",
                                        "pivot": {
                                            "user_id": 1,
                                            "file_id": 1
                                        }
                                    },
                                    {
                                        "id": 2,
                                        "name": "jx3m8NPBgy.png",
                                        "type": "image",
                                        "path": "public/image/jx3m8NPBgy.png",
                                        "status": 0,
                                        "created_at": "2022-09-29T17:03:15.000000Z",
                                        "updated_at": "2022-09-29T17:13:16.000000Z",
                                        "deleted_at": null,
                                        "url": "http://domain.com/storage/image/jx3m8NPBgy.png",
                                        "pivot": {
                                            "user_id": 1,
                                            "file_id": 2
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "getdata fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/storage/client/file/get": {
            "post": {
                "tags": [
                    "Storage Api file"
                ],
                "summary": "get file",
                "description": "use token",
                "operationId": "get file",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "path": {
                                        "description": "file path to save",
                                        "type": "string",
                                        "example": "Admin/image/redis.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/FileSystem"
                                        }
                                    ]
                                },
                                "example": {
                                    "id": 1,
                                    "name": "jx3m8NPBgy.png",
                                    "type": "image",
                                    "path": "public/image/jx3m8NPBgy.png",
                                    "status": 0,
                                    "created_at": "2022-09-29T17:03:15.000000Z",
                                    "updated_at": "2022-09-29T17:13:16.000000Z",
                                    "deleted_at": null,
                                    "url": "http://domain.com/storage/image/jx3m8NPBgy.png",
                                    "pivot": {
                                        "user_id": 1,
                                        "file_id": 1
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "notfound"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/storage/client/file/upload": {
            "post": {
                "tags": [
                    "Storage Api file"
                ],
                "summary": "upload file.",
                "description": "upload File",
                "operationId": "9ab25748707e4aa63e5e5a6760f76e34",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "path": {
                                        "description": "file path to save",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/FileSystem"
                                        }
                                    ]
                                },
                                "example": {
                                    "message": "upload success",
                                    "file": {
                                        "id": 1,
                                        "name": "jx3m8NPBgy.png",
                                        "type": "image",
                                        "path": "public/image/jx3m8NPBgy.png",
                                        "status": 0,
                                        "created_at": "2022-09-29T17:03:15.000000Z",
                                        "updated_at": "2022-09-29T17:13:16.000000Z",
                                        "deleted_at": null,
                                        "url": "http://domain.com/storage/image/jx3m8NPBgy.png"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "upload fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/storage/client/file/update/{id}": {
            "post": {
                "tags": [
                    "Storage Api file"
                ],
                "summary": "update file.",
                "description": "update File",
                "operationId": "180164f5da329e936e67e3deb2d98560",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id file",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/FileSystem"
                                        }
                                    ]
                                },
                                "example": {
                                    "message": "update success",
                                    "file": {
                                        "id": 1,
                                        "name": "jx3m8NPBgy.png",
                                        "type": "image",
                                        "path": "public/image/jx3m8NPBgy.png",
                                        "status": 0,
                                        "created_at": "2022-09-29T17:03:15.000000Z",
                                        "updated_at": "2022-09-29T17:13:16.000000Z",
                                        "deleted_at": null,
                                        "url": "http://domain.com/storage/image/jx3m8NPBgy.png"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "update fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/storage/client/file/delete/{id}": {
            "delete": {
                "tags": [
                    "Storage Api file"
                ],
                "summary": "delete file",
                "description": "use token",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id file",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "delete success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "delete fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "FileSystem": {
                "title": "FileSystem",
                "description": "FileSystem model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "file_name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "pivot": {
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "FileSystem"
                }
            },
...
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8001/api/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8001/api/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8001/api/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "passport": []
        }
    ]
}